Fix StorageAccessPointLostAndFound not loading correctly
It references the StorageBase in the vanilla scene, so when we move it to our scene it loses that reference. This is a little hacky for now, perhaps in the future we could add a script to all objects that reference a StorageBase and assign them the correct one once the scene has been loaded.