Refactor map loading to add loading screen messages
Before, all map assets were loading before WorldStreamingInit#Awake was called. This mean the game would just have a black screen and be unresponsive while loading.

This changes waits until the loading screen has been initialized to start loading the map. To keep things (somewhat) organized, the map loading logic has been moved to the MapLoader class and into a coroutine so we can update the loading screen between blocking actions.

To keep the proper order of execution, WorldStreamingInit#Awake is now skipped, and we manually run the LoadingRoutine coroutine once we've finished. It's OK that the base Awake call gets skipped since it'll be initialized when our WaitForLoadingScreen requests it.