Refactor the map selector setting
The setting is no longer backed by an index from an OrderedDictionary. It's now based on the name of the map. This means the map that loads will no longer change if other maps are added/removed.

This also fixes a bug where savegame backups weren't using the proper name.