Save BasicMapInfo to mapInfo.json instead of the full MapInfo, add version field
Since MapInfo contains some large data sets (like serialized images) it's not ideal to write all that to json when all we read from it is the name.

Because of this, the BasicMapInfo class has been moved to the MapifyEditor module, and it's now used for serialization to that file. This also means the variable names don't have to match MapInfo's

Since it was convenient, the Mapify version the map was built with is also included for future use.