Improve log handling
The public Logger property has been replaced with LogX functions, that accept object's instead of string, making debugging easier.

With that, a #LogDebug method has been added, along with a setting to toggle whether debugs are logged.

This allows us to be more verbose in code, but not clutter people's log files unless we're trying to figure out an issue.