Add initial support for Simulator
This updates the code to compile against the latest closed beta for Simulator. It's still quite far from functional, however, the map selection screen and localization support is mostly finished.

The logic in most patches will need to be updated to see if the loaded map is custom or not, since it currently assumes that it is due to patches only being applied for custom maps.

A project restructure should be done before progressing further, as putting all patches together makes it extremely hard to tell what individual pieces go together.